DeepLearningSchool/homework_CNN_Deploy_in_heroku/requirements.txt
gunicorn
https://download.pytorch.org/whl/cpu/torch-1.1.0-cp37-cp37m-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.3.0-cp37-cp37m-linux_x86_64.whl
Pillow
opencv-python
matplotlib
flask
flask_dropzone